Use wxArrayInt64 and fix typo

+++ src/filtertrans.h
@@ -40,7 +40,7 @@ public:
     template<class MODEL, class DATA = typename MODEL::Data>
     bool checkCategory(const DATA& tran, const std::map<int64, typename MODEL::Split_Data_Set> & splits);
     bool mmIsRecordMatches(const Model_Checking::Data &tran
-        , const std::map<int, Model_Splittransaction::Data_Set>& split);
+        , const std::map<int64, Model_Splittransaction::Data_Set>& split);
 
     wxString getHTML();
